+++ index.d.ts
@@ -1,19 +1,14 @@
-export = UserAgentBag;
-export as namespace UserAgentBag;
-
-declare namespace UserAgentBag {
-  export type Entry = [string, string | null];
-  export type Entries = Iterable<Entry>;
-}
-
 declare class UserAgentBag {
-  constructor(arg?: string | UserAgentBag.Entries | null);
+  constructor(arg?: null | string | Array<Entry> | Iterable<Entry>);
 
-  get(product: string): string | null | void;
-  getAll(product: string): Array<string | null | void>;
+  entries(): Iterable<Entry>;
+  get(product: string): undefined | null | string;
+  getAll(product: string): Array<undefined | null | string>;
   has(product: string): boolean;
-  entries(): UserAgentBag.Entries;
   size(): number;
-
   toString(): string;
 }
+
+type Entry = [string, null | string];
+
+export = UserAgentBag;